Impl 90% of API methods, add base test in main
Implement most of the methods exposed as an API, with the exception of the
method rcSubmt(), which will be implemented in the next iterations.
Hence, as of this point, the software is still not complete.

Also add very basic test in main for the creation of a device and its
destruction.