Start impl. of opaque objs, refactor debug output
Start implementing the opaque objects design for the API, so that the
memory for buffers is all managed by the user of the API.
Implemented for Devices; will be implemented for the other objects.

Refactor how the address of ptrs is logged, using %p.

Add files to the .gitignore.

Update submodules.